Change format stream: push to sink stream
Previously the sink (e.g. http ice) output stream pulled bytes on demand
from a lazy seq given to it by the format stream. This won't work when
trying to sleep in real time. Instead the format stream now pushes byte
arrays to the sink stream. This may still not work well if and when it
takes longer to build the byte arrays than is required to sleep. This
excess time should be amortized over the stream.